+++ tests/rock_paper_scissors/keywords.py
@@ -1,3 +1,7 @@
+# State machines
+
+MACHINE_A = "[[MachineA]]"
+
 # States
 INIT = "[Init]"
 GAME = "[Game]"