+++ tests/rock_paper_scissors/layer1_transition_map/game.py
@@ -1,5 +1,7 @@
 import re
 from state_machine_py.abstract_state import AbstractState
+from state_machine_py.intermachine import Intermachine
+from state_machine_py.multiple_state_machine import MultipleStateMachine
 from state_machine_py.request import Request
 from tests.rock_paper_scissors.context import Context
 from tests.rock_paper_scissors.keywords import E_DRAW, E_LOOPBACK, E_LOSE, E_WIN, GAME
@@ -38,7 +40,7 @@ class GameState(AbstractState):
         """
         super().exit(req)
 
-        matched = self._janken_pattern.match(req.line)
+        matched = self._janken_pattern.match(req.intermachine.dequeue_myself())
         if matched:
             janken = matched.group(1)
 
@@ -72,35 +74,36 @@ class GameState(AbstractState):
 # Test
 # python.exe -m layer1_transition_map.game
 if __name__ == "__main__":
+    intermachine = Intermachine(MultipleStateMachine(), "[[TestMachine]]")
     context = Context()
     state = GameState()
 
-    line = 'R'
-    req = Request(context, line, [])
+    req = Request(context, intermachine)
+    req.intermachine.enqueue_myself('R')
     edge = state.exit(req)
     if edge == E_LOSE:
         print('.', end='')
     else:
         print('f', end='')
 
-    line = 'S'
-    req = Request(context, line, [])
+    req = Request(context, intermachine)
+    req.intermachine.enqueue_myself('S')
     edge = state.exit(req)
     if edge == E_WIN:
         print('.', end='')
     else:
         print('f', end='')
 
-    line = 'P'
-    req = Request(context, line, [])
+    req = Request(context, intermachine)
+    req.intermachine.enqueue_myself('P')
     edge = state.exit(req)
     if edge == E_DRAW:
         print('.', end='')
     else:
         print('f', end='')
 
-    line = 'W'
-    req = Request(context, line, [])
+    req = Request(context, intermachine)
+    req.intermachine.enqueue_myself('W')
     edge = state.exit(req)
     if edge == E_LOOPBACK:
         print('.', end='')